# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Greater Bayelsa leadership development platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List meeting records within scope
      parameters:
        - name: villageId
          in: query
          required: false
          schema: { type: integer }
      responses:
        "200":
          description: Meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: submitMeeting
      tags: [meetings]
      summary: Submit a monthly meeting update (village head)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingSubmission"
      responses:
        "201":
          description: Meeting recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get a meeting with media and attendance
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Meeting detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------- Attendance ----------
  /attendance/checkin:
    post:
      operationId: attendanceCheckin
      tags: [attendance]
      summary: QR self-check-in for a meeting (member scans code)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinBody"
      responses:
        "201":
          description: Attendance logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
        "404":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /attendance/video:
    post:
      operationId: attendanceVideo
      tags: [attendance]
      summary: Record video-verified attendance for a member (leader)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VideoAttendanceBody"
      responses:
        "201":
          description: Attendance logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # ---------- HQ requests ----------
  /hq-requests:
    get:
      operationId: listHqRequests
      tags: [hq-requests]
      summary: List requests to headquarters within scope
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HqRequest"
  /hq-requests/{id}:
    patch:
      operationId: updateHqRequest
      tags: [hq-requests]
      summary: Respond to / resolve an HQ request (HQ)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HqRequestUpdate"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HqRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------- Calendar ----------
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: Calendar events visible to the caller (org-wide + local)
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (HQ = org-wide; leaders = local)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (creator or HQ)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ---------- Public posts ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Meeting:
      type: object
      required:
        [id, villageId, heldOn, attendanceCount, discussionPoints, createdAt]
      properties:
        id: { type: integer }
        villageId: { type: integer }
        villageName: { type: [string, "null"] }
        heldOn: { type: string }
        attendanceCount: { type: integer }
        unitBreakdown:
          type: [array, "null"]
          items:
            type: object
            required: [unitId, count]
            properties:
              unitId: { type: integer }
              count: { type: integer }
        discussionPoints: { type: string }
        submittedById: { type: integer }
        checkinCode: { type: [string, "null"] }
        verifiedAttendance: { type: [integer, "null"] }
        createdAt: { type: string }
    MeetingAgendaSegments:
      description: "Fixed six-segment monthly meeting agenda. All six keys are required; values must be non-empty strings."
      type: object
      required: [opening, wellbeing, updates, openFloor, dues, closing]
      properties:
        opening: { type: string, minLength: 1 }
        wellbeing: { type: string, minLength: 1 }
        updates: { type: string, minLength: 1 }
        openFloor: { type: string, minLength: 1 }
        dues: { type: string, minLength: 1 }
        closing: { type: string, minLength: 1 }
    MeetingSubmission:
      type: object
      required: [heldOn, attendanceCount, agenda, photos, video]
      properties:
        heldOn: { type: string }
        attendanceCount: { type: integer, minimum: 0 }
        unitBreakdown:
          type: [array, "null"]
          items:
            type: object
            required: [unitId, count]
            properties:
              unitId: { type: integer }
              count: { type: integer }
        agenda:
          $ref: "#/components/schemas/MeetingAgendaSegments"
        hqRequest: { type: [string, "null"] }
        photos:
          type: array
          minItems: 5
          items: { type: string }
        targetVillageId:
          type: [integer, "null"]
          description: "HQ only: specify the village this meeting belongs to"
        video:
          type: object
          required: [objectPath, durationSeconds]
          properties:
            objectPath: { type: string }
            durationSeconds: { type: integer, minimum: 120 }
    MeetingMedia:
      type: object
      required: [id, kind, objectPath]
      properties:
        id: { type: integer }
        kind: { type: string }
        objectPath: { type: string }
        durationSeconds: { type: [integer, "null"] }
    MeetingDetail:
      allOf:
        - $ref: "#/components/schemas/Meeting"
        - type: object
          required: [media, attendance]
          properties:
            media:
              type: array
              items:
                $ref: "#/components/schemas/MeetingMedia"
            attendance:
              type: array
              items:
                $ref: "#/components/schemas/AttendanceRecord"
    AttendanceRecord:
      type: object
      required: [id, meetingId, userId, method, createdAt]
      properties:
        id: { type: integer }
        meetingId: { type: integer }
        userId: { type: integer }
        memberName: { type: [string, "null"] }
        method: { type: string }
        videoObjectPath: { type: [string, "null"] }
        createdAt: { type: string }
    CheckinBody:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 4 }
    VideoAttendanceBody:
      type: object
      required: [meetingId, userId, videoObjectPath]
      properties:
        meetingId: { type: integer }
        userId: { type: integer }
        videoObjectPath: { type: string }
    HqRequest:
      type: object
      required: [id, villageId, body, status, createdAt]
      properties:
        id: { type: integer }
        meetingId: { type: [integer, "null"] }
        villageId: { type: integer }
        villageName: { type: [string, "null"] }
        submittedById: { type: integer }
        submitterName: { type: [string, "null"] }
        body: { type: string }
        status: { type: string }
        response: { type: [string, "null"] }
        createdAt: { type: string }
    HqRequestUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [open, responded, resolved] }
        response: { type: [string, "null"] }
    Event:
      type: object
      required: [id, title, startsAt, scope, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: [string, "null"] }
        location: { type: [string, "null"] }
        startsAt: { type: string }
        scope: { type: string }
        villageId: { type: [integer, "null"] }
        unitId: { type: [integer, "null"] }
        createdById: { type: [integer, "null"] }
        createdAt: { type: string }
    EventInput:
      type: object
      required: [title, startsAt, scope]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: [string, "null"] }
        location: { type: [string, "null"] }
        startsAt: { type: string }
        scope: { type: string, enum: [org, village, unit] }
        villageId: { type: [integer, "null"] }
        unitId: { type: [integer, "null"] }
...