# lib/api-zod/src/generated/types/checkinBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Greater Bayelsa leadership development platform API
 * OpenAPI spec version: 0.2.0
 */

export interface CheckinBody {
  /** @minLength 4 */
  code: string;
}